requirements.txt
antlr4-python3-runtime~=4.13.2
fastapi
uvicorn
pydantic
pytest
pandas
fastapi-utils
typing-inspect